Fixed int type bug

+++ nagi/neat.py
@@ -392,7 +392,7 @@ class Population(object):
             species_size = assigned_number_of_offspring_per_species[species_id]
             old_members = sorted(species.members, key=lambda x: fitnesses[x.key], reverse=True)
 
-            for genome in old_members[:max(species_size*ELITISM, MIN_SPECIES_SIZE)]:
+            for genome in old_members[:max(int(species_size*ELITISM), MIN_SPECIES_SIZE)]:
                 new_population_of_genomes[genome.key] = genome
                 species_size -= 1
